refactor: faculty, department を分割
