+++ workspaces/schema/src/entity/user.ts
@@ -2,7 +2,8 @@ import * as v from "valibot";
 import { MaxLines } from "../common/max-lines";
 import { RHFableArray } from "../common/rhf-array";
 import { UserCertification } from "./certification";
-import { DepartmentId, FacultyId } from "./department";
+import { DepartmentId } from "./department";
+import { FacultyId } from "./faculty";
 import { GradeId } from "./grade";
 import { ExportableOAuthConnection } from "./oauth-internal/oauth-connection";
 import { Role } from "./role";
